\name{00.numDeriv.Intro}
\alias{00.numDeriv.Intro}
\alias{numDeriv.Intro}
\docType{package}

\title{Accurate Numerical Derivatives}

\description{Calculate (accurate) numerical approximations to derivatives.}

\details{
\tabular{ll}{
Package: \tab numDeriv\cr
Depends: \tab R (>= 2.2.1)\cr
License: \tab GPL Version 2.\cr
}
The main functions are
\code{\link{grad}}, \code{\link{jacobian}}, \code{\link{hessian}}
and \code{\link{genD}}. 

\author{Paul Gilbert, others}

Maintainer: Paul Gilbert <pgilbert@bank-banque-canada.ca>  
}
\references{
   Linfield, G. R. and Penny, J. E. T. (1989) \emph{Microcomputers in Numerical 
   Analysis}. New York: Halsted Press.

  Fornberg and Sloan, (1994) \emph{Acta Numerica}, p. 203-267; Table 1, page 213)
}
\keyword{package}

