% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_single_coverage_heatmap.R
\name{add_coverage_heatmap}
\alias{add_coverage_heatmap}
\title{add_coverage_heatmap}
\usage{
add_coverage_heatmap(p, mat, x = iHeatmapR:::default_x(mat),
  groups = p$row_groups, signal = NULL, name = "Signal", summary = TRUE,
  scale = c("rows", "none"), scale_method = c("normalize", "standardize",
  "center"), ...)
}
\arguments{
\item{mat}{coverage matrix}

\item{x}{x axis labels}

\item{groups}{pre-determined groups for rows}

\item{signal}{signal along row}

\item{name}{name of colorbar}

\item{summary}{make summary plot, boolean}

\item{scale}{"rows" or "none"}

\item{scale_method}{"normalize", "standardize", "center"}
}
\value{
iHeatmap object
}
\description{
add_coverage_heatmap
}
\author{
Alicia Schep
}

