% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_single_coverage_heatmap.R
\name{add_multi_coverage_heatmap}
\alias{add_multi_coverage_heatmap}
\title{add_multi_coverage_heatmap}
\usage{
add_multi_coverage_heatmap(p, mats, x = iHeatmapR:::default_x(mats[[1]]),
  groups = NULL, signal = lapply(mats, rowMeans, na.rm = TRUE),
  plot_signal = TRUE, name = "Coverage", signal_name = "Avg. Coverage",
  summary = TRUE, source = "HM", scale_method = c("localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar", "none"), pct = 0.95,
  scale_factor = rep(1, length(mats)), share_z = TRUE, ticktext = TRUE,
  start = x[1], end = default_end(x), xlab = "Position",
  font = list(size = 8), ...)
}
\arguments{
\item{x}{x axis labels}

\item{groups}{pre-determined groups for rows}

\item{signal}{signal along row}

\item{name}{name of colorbar}

\item{summary}{make summary plot, boolean}

\item{source}{source name in plotly}

\item{scale_method}{method to use for scaling}

\item{mat}{coverage matrix}

\item{y}{y axis labels}

\item{scale}{scale rows? or none}
}
\value{
iHeatmap object
}
\description{
add_multi_coverage_heatmap
}
\author{
Alicia Schep
}

