% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_ly.R
\name{add_tx_shapes}
\alias{add_tx_shapes}
\title{Title}
\usage{
add_tx_shapes(plotly_obj, tx_info, target_range, grt_ax, y_scaling = 0)
}
\arguments{
\item{plotly_obj}{}

\item{tx_info}{}

\item{target_range}{}

\item{grt_ax}{}
}
\description{
Title
}

