% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{addExpressionIdentifier}
\alias{addExpressionIdentifier}
\title{Title}
\usage{
addExpressionIdentifier(file_info, es_identifier)
}
\arguments{
\item{file_info}{}

\item{es_identifier}{}
}
\description{
Title
}

