\name{addTA}
\alias{addTA}
\alias{addVo}
\alias{addMA}
\alias{addSMI}
\alias{addExpiry}

\title{ Add Technical Indicator to Chart } 
\description{
Add dynamically or via TA argument specified technical indicator
to chart.
}
\usage{
addVo()
}
%- maybe also 'usage' for other objects documented here.
\details{
Add TA to chart
}
\value{
Called for its side effects, an object to class \code{chobTA}
will be returned.
}
\references{ TTR package }
\author{ Jeffrey A. Ryan }
\keyword{ utilities }
