% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_profile.R
\name{aggregate_profile_plot}
\alias{aggregate_profile_plot}
\title{aggregate_profile_plot}
\usage{
aggregate_profile_plot(mats, groups = NULL, colors = "Dark2",
  positions = colnames(mats[[1]]), ylab = "ChIP Signal",
  xlab = "Position", scale_method = c("none", "localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar"), pct = 0.95,
  scale_factor = 1, showlegend = TRUE)
}
\description{
aggregate_profile_plot
}
\author{
Alicia Schep
}

