\name{AIC.ALKmodel}
\alias{AIC.ALKmodel}
\title{Compute Akaike Information Criterion for an age-length key model.}
\usage{
  AIC.ALKmodel(x, k = 2)
}
\arguments{
  \item{x}{an ALKmodel object.}

  \item{k}{Penalty for number of parameters}
}
\value{
  AIC value.
}
\description{
  Compute Akaike Information Criterion for an age-length
  key model.
}

