\name{AIC}
\alias{AIC}
\alias{AIC,depmix-method}
\alias{BIC}
\alias{BIC,depmix-method}

\title{Compute AIC and BIC for depmix objects}
\description{
Compute AIC and BIC for depmix objects.
}
\usage{
AIC(object, ..., k = 2)
BIC(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{A \code{depmix} model.}
  \item{\dots}{Not used currently.}
  \item{k}{The penalty factor which defaults to 2.}
}
\value{
	The value of the AIC and BIC
}
\references{}
\author{Ingmar Visser}
\keyword{htest}
