\name{anova.ALKmodel}
\alias{anova.ALKmodel}
\title{Compute approximate likelihood ratio test for the equality of two age-length keys.}
\usage{
  anova.ALKmodel(object, object2, ...)
}
\arguments{
  \item{object}{An object of class 'ALKmodel'}

  \item{object2}{An object of class 'ALKmodel'}
}
\value{
  a p-value.
}
\description{
  Compute approximate likelihood ratio test for the
  equality of two age-length keys.
}

