\name{arg.names}
\alias{arg.names}

\title{
  Argument Names
}

\description{
  Returns the argument names of a function or call.
}

\usage{
  arg.names(x)
}

\arguments{
  \item{x}{a function, call, or a character vector of length 1 containing the name of a function.}
}

\value{
  a character vector containing the names of the arguments.
}

\examples{
  arg.names(lm)
}

\keyword{utilities}
