% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{avg_peak_in_range}
\alias{avg_peak_in_range}
\title{Title}
\usage{
avg_peak_in_range(gr, peaks_gr)
}
\arguments{
\item{gr}{GRanges to find peaks in}

\item{peaks_gr}{GRanges of Macs peaks, e.g. from readMacsPeaks. mcols must include summit and pileup}
}
\description{
Title
}
