% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_ly.R
\name{browserly_tx_track}
\alias{browserly_tx_track}
\title{browserly_tx_track
Make the invisible points for an annotation track.}
\usage{
browserly_tx_track(tx_info, track_name, add_shapes = FALSE,
  target_range = NULL)
}
\arguments{
\item{tx_info}{GRanges: mcols describe the annotation features}
}
\value{
plotly_object
}
\description{
Currently this is only a helper function, but it could be made to stand alone
}

