\name{bss}
\alias{bss}
\docType{data}
\title{
Benevolent Sexism Scale
}
\description{
Responses by 201 undergraduates at the University of Illinois at Urbana-Champaign to Spence, Helmreich, and Stapp's (1974) 15-item Attitudes Toward Women Scale.
}
\usage{data(bss)}
\format{
  A data frame with 201 observations on the following 16 variables.
  \describe{
    \item{\code{subj}}{Subject identification number.}
    \item{\code{bss01}}{No matter how accomplished he is, a man is not truly complete as a person unless he has the love of a woman.}
    \item{\code{bss02}}{In a disaster, women ought not necessarily to be rescued before men.}
    \item{\code{bss03}}{People are often truly happy in life without being romantically involved with a member of the other sex.}
    \item{\code{bss04}}{Many women have a quality of purity that few men possess.}
    \item{\code{bss05}}{Women should be cherished and protected by men.}
    \item{\code{bss06}}{Every man ought to have a woman whom he adores.}
    \item{\code{bss07}}{Men are complete without women.}
    \item{\code{bss08}}{A good woman should be set on a pedestal by her man.}
    \item{\code{bss09}}{Women, compared to men, tend to have a superior moral sensibility.}
    \item{\code{bss10}}{Men should be willing to sacrifice their own well being in order to provide financially for the women in their lives.}
    \item{\code{bss11}}{Women, as compared to men, tend to have a more refined sense of culture and good taste.}
  }
}
\details{
Administered in the fall of 1999. Participants responded to each item using a 1-7 Likert scale in which 7 indicated \dQuote{Strongly Agree} and 1 indicated \dQuote{Strongly Disagree}.
}

\references{
Spence, J. T., Helmreich, R., & Stapp, J. (1975). Ratings of self and peers on sex role attributes and their relation to self-esteem and conceptions of masculinity and femininity. Journal of Personality and Social Psychology, 32(1), 29-39.
}
\examples{
data(bss)
}
\keyword{datasets}
