% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_ly.R
\name{calc_title_margin}
\alias{calc_title_margin}
\title{calc_title_margin
Calculate how many pixels of space the titles need. Assumes the titles are horizontal}
\usage{
calc_title_margin(titles, padding = 0, letter_width = 9)
}
\arguments{
\item{titles}{vector}

\item{padding}{int: number of character to pad to the maximium title found}

\item{letter_width}{int: number of pixels each letter requires}
}
\value{

}
\description{
calc_title_margin
Calculate how many pixels of space the titles need. Assumes the titles are horizontal
}

