\name{census}
\alias{census}
\alias{census.assMatrix}
\docType{data}
\title{Census data set}
\description{

}
\usage{data(census)
data(census.assMatrix)}
\format{
  A data frame with 48842 observations on the following 14 variables.
  \describe{
    \item{age}{a factor with levels \code{middle-aged} \code{old} \code{senior} \code{young}}
    \item{workclass}{a factor with levels \code{Federal-gov} \code{Local-gov} \code{Never-worked} \code{Private} \code{Self-emp-inc} \code{Self-emp-not-inc} \code{State-gov} \code{Without-pay}}
    \item{education}{a factor with levels \code{10th} \code{11th} \code{12th} \code{1st-4th} \code{5th-6th} \code{7th-8th} \code{9th} \code{Assoc-acdm} \code{Assoc-voc} \code{Bachelors} \code{Doctorate} \code{HS-grad} \code{Masters} \code{Preschool} \code{Prof-school} \code{Some-college}}
    \item{education-num}{a factor with levels \code{1} \code{10} \code{11} \code{12} \code{13} \code{14} \code{15} \code{16} \code{2} \code{3} \code{4} \code{5} \code{6} \code{7} \code{8} \code{9}}
    \item{marital-status}{a factor with levels \code{Divorced} \code{Married-AF-spouse} \code{Married-civ-spouse} \code{Married-spouse-absent} \code{Never-married} \code{Separated} \code{Widowed}}
    \item{occupation}{a factor with levels \code{Adm-clerical} \code{Armed-Forces} \code{Craft-repair} \code{Exec-managerial} \code{Farming-fishing} \code{Handlers-cleaners} \code{Machine-op-inspct} \code{Other-service} \code{Priv-house-serv} \code{Prof-specialty} \code{Protective-serv} \code{Sales} \code{Tech-support} \code{Transport-moving}}
    \item{relationship}{a factor with levels \code{Husband} \code{Not-in-family} \code{Other-relative} \code{Own-child} \code{Unmarried} \code{Wife}}
    \item{race}{a factor with levels \code{Amer-Indian-Eskimo} \code{Asian-Pac-Islander} \code{Black} \code{Other} \code{White}}
    \item{sex}{a factor with levels \code{Female} \code{Male}}
    \item{capital-gain}{a factor with levels \code{high} \code{medium} \code{none} \code{small}}
    \item{capital-loss}{a factor with levels \code{medium} \code{none} \code{small}}
    \item{hours-per-week}{a factor with levels \code{full-time} \code{half-time} \code{overtime} \code{too-many}}
    \item{native-country}{a factor with levels \code{Cambodia} \code{Canada} \code{China} \code{Columbia} \code{Cuba} \code{Dominican-Republic} \code{Ecuador} \code{El-Salvador} \code{England} \code{France} \code{Germany} \code{Greece} \code{Guatemala} \code{Haiti} \code{Holand-Netherlands} \code{Honduras} \code{Hong} \code{Hungary} \code{India} \code{Iran} \code{Ireland} \code{Italy} \code{Jamaica} \code{Japan} \code{Laos} \code{Mexico} \code{Nicaragua} \code{Outlying-US(Guam-USVI-etc)} \code{Peru} \code{Philippines} \code{Poland} \code{Portugal} \code{Puerto-Rico} \code{Scotland} \code{South} \code{Taiwan} \code{Thailand} \code{Trinadad&Tobago} \code{United-States} \code{Vietnam} \code{Yugoslavia}}
    \item{salary}{a factor with levels \code{small} \code{large}}
  }
}
\details{
  
}
\source{http://www.ics.uci.edu/~mlearn/MLRepository.html
}
\references{
  Blake, C.L. \& Merz, C.J. (1998).
  UCI Repository of machine learning databases
  [http://www.ics.uci.edu/~mlearn/MLRepository.html].
  Irvine, CA: University of California, Department of Information and Computer Science.
}
\keyword{datasets}
