% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RIGHT.R
\name{clean}
\alias{clean}
\title{Cleanup RIGHT Object}
\usage{
clean(obj)
}
\arguments{
\item{obj}{RIGHT object.}
}
\description{
Cleanup RIGHT Object
}
\examples{
\dontrun{
obj <- RIGHT(plot(conc ~ Time, Theoph), Theoph)
clean(obj)
}
}

