% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_heatmap.R
\name{cluster_heatmap}
\alias{cluster_heatmap}
\title{cluster_heatmap}
\usage{
cluster_heatmap(mat, x = chipVis:::default_x(mat),
  y = chipVis:::default_y(mat), row_order = c("none", "hclust", "kmeans",
  "groups"), col_order = c("none", "hclust", "kmeans", "groups"),
  row_groups = NULL, col_groups = NULL, row_anno = NULL,
  col_anno = NULL, row_k = NULL, col_k = NULL,
  row_clust_dist = stats::dist, col_clust_dist = stats::dist,
  name = "Signal", source = "HM", scale = TRUE)
}
\arguments{
\item{mat}{matrix}

\item{x}{x axis labels}

\item{y}{y axis labels}

\item{row_order}{row order method}

\item{col_order}{col order method}

\item{row_groups}{pre-determined groups for rows}

\item{col_groups}{pre-determined groups for col}

\item{row_k}{k to use for kmeans clustering or cutting heirarchical clustering}

\item{col_k}{k to use for kmeans clustering or cutting heirarchical clustering}

\item{row_clust_dist}{distance function to uses for row clustering}

\item{col_clust_dist}{distance function to uses for col clustering}

\item{name}{name of colorbar}

\item{source}{source name in plotly}
}
\value{
list with five components, 1) plot -- function to plot
2) row_order -- row order used, 3) dendro hclust object if heirarchical clustering
performed for rows  4) col_order -- row order used, 5) dendro hclust object if 
heirarchical clustering
performed for cols
}
\description{
cluster_heatmap
}
\author{
Alicia Schep
}

