\name{core}
\alias{core}
\alias{core<-}
\title{Extract core from object or assign core to object}
\description{Extract core from objects.  Particularly time series objects.}
\usage{
core(x)
core(x) <- value
}
\arguments{
  \item{x}{ an object which contains core }
  \item{value}{ new core to be assigned to object }
  \item{\dots}{ further arguments to function }
}
\value{
  a vector of core
}
\author{ Whit Armstrong }
\keyword{ts}
