% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggplot_RIGHT.R
\name{createGgplot}
\alias{createGgplot}
\title{Extract data object's name}
\usage{
createGgplot(data, ...)
}
\arguments{
\item{data}{a data.frame object.}

\item{...}{other options to draw graph(colour, fill)}
}
\description{
Function to create data object name in ggplot function.
}
\examples{
\dontrun{
RIGHT(print(ggplot(Theoph, aes(Time, conc, colour=Subject)) + geom_point()))
}
}
\seealso{
\code{\link{ggplot2}}
}

