\name{DATRAS-internal}
\alias{$.DATRASraw}
\alias{$<-.DATRASraw}
\alias{addExtraVariables}
\alias{addHaulID}
\alias{addSpectrum}
\alias{checkSpectrum}
\alias{DATRAS-internal}
\alias{fixMissingHaulIds}
\alias{getAccuracyCM}
\alias{length.DATRASraw}
\alias{print.DATRASraw}
\alias{readICES}
\alias{refactorHaulLevels}
\alias{reorderTimeLevels}
\alias{summary.DATRASraw}
\alias{tryCatch.W.E}
\alias{xtraVars}
\title{Internal DATRAS Functions}
\description{
  Internal DATRAS functions
}
\details{
  These are not to be called by the user (or in some cases
  are just waiting for proper documentation to be written
  :).
}

