% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{dcolorscale}
\alias{dcolorscale}
\title{dcolorscale}
\usage{
dcolorscale(x = 2, palette = "Dark2", cols = NULL)
}
\arguments{
\item{x}{number of items}

\item{palette}{name of RColorBrewer palette}

\item{cols}{vector of color names, overrides palette if given}
}
\value{
data.frame of colors and breaks to give as colorscale input to plotly
}
\description{
dcolorscale
}
\author{
Alicia Schep
}

