\name{exampleFunction}
\alias{exampleFunction}
\title{A Simple Example Function}
\usage{
  exampleFunction(x, y)
}
\arguments{
  \item{x}{integer}

  \item{y}{integer}
}
\value{
  integer
}
\description{
  .. content for the "Description" section of the
  documentation ..
}
\details{
  .. content for "Details" section of the documentation ..
}
\author{
  Cory Barr
}

