% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChipseq_functions.R
\name{expandPairedTable}
\alias{expandPairedTable}
\title{expand the paired sample table with default directory structure}
\usage{
expandPairedTable(dat.pair, order = "sample.first")
}
\arguments{
\item{dat.pair}{input paired sample table}

\item{order}{sample.first means the top directory will be named by the sample
pair, subdir names are macs_output etc.}
}
\value{
data frame of expanded sample table
}
\description{
expand the paired sample table with default directory structure, adding directories for
macs output, peak summary, and peak motif. Directories predifined by the user will
be preserved
}
\author{
Jinfeng Liu
}

