% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackview_functions.R
\name{extend_grange}
\alias{extend_grange}
\title{Extend GRange}
\usage{
extend_grange(gr, extend = 0)
}
\arguments{
\item{gr}{GRange}

\item{extend}{scalar: 1 value if symmetric extension. Otherwise in order start, end}
}
\value{
gr GRange
}
\description{
extend the GRange object in either direction
}
\author{
Justin Finkle
}

