\name{extract-methods}
\docType{methods}
\alias{extract-methods}
\alias{extract,clm-method}
\alias{extract,clogit-method}
\alias{extract,coxph-method}
\alias{extract,coxph.penal-method}
\alias{extract,ergm-method}
\alias{extract,gee-method}
\alias{extract,glm-method}
\alias{extract,glmerMod-method}
\alias{extract,gls-method}
\alias{extract,gmm-method}
\alias{extract,lm-method}
\alias{extract,lme-method}
\alias{extract,lmerMod-method}
\alias{extract,lmrob-method}
\alias{extract,lnam-method}
\alias{extract,lrm-method}
\alias{extract,mer-method}
\alias{extract,negbin-method}
\alias{extract,nlmerMod-method}
\alias{extract,plm-method}
\alias{extract,pmg-method}
\alias{extract,polr-method}
\alias{extract,Relogit-method}
\alias{extract,rlm-method}
\alias{extract,rq-method}
\alias{extract,sclm-method}
\alias{extract,simex-method}
\alias{extract,stergm-method}
\alias{extract,svyglm-method}
\alias{extract,systemfit-method}
\alias{extract,tobit-method}
\title{Methods for Function \code{extract} in Package \pkg{texreg}}
\description{
Methods for function \code{extract} in package \pkg{texreg}.
}
\section{Methods}{
\describe{

\item{\code{extract.clm}}{
An extract function for clm objects from the \pkg{ordinal} package.
}

\item{\code{extract.clogit}}{
An extract function for clogit objects from the \pkg{survival} package.
}

\item{\code{extract.coxph}}{
An extract function for coxph objects from the \pkg{survival} package.
}

\item{\code{extract.coxph.penal}}{
An extract function for coxph.penal objects from the \pkg{survival} package.
}

\item{\code{extract.ergm}}{
An extract function for ergm objects from the \pkg{ergm} package.
}

\item{\code{extract.gee}}{
An extract function for gee objects from the \pkg{gee} package.
}

\item{\code{extract.glm}}{
An extract function for glm objects from the \pkg{stats} package.
}

\item{\code{extract.glmerMod}}{
An extract function for glmerMod objects from the (new) \pkg{lme4} package.
}

\item{\code{extract.gls}}{
An extract function for gls objects from the \pkg{nlme} package.
}

\item{\code{extract.gmm}}{
An extract function for gmm objects from the \pkg{gmm} package.
}

\item{\code{extract.lm}}{
An extract function for lm objects from the \pkg{stats} package.
}

\item{\code{extract.lme}}{
An extract function for lme objects from the \pkg{nlme} package.
}

\item{\code{extract.lmerMod}}{
An extract function for lmerMod objects from the (new) \pkg{lme4} package.
}

\item{\code{extract.lmrob}}{
An extract function for lmrob objects from the \pkg{robustbase} package.
}

\item{\code{extract.lnam}}{
An extract function for lnam objects from the \pkg{sna} package.
}

\item{\code{extract.lrm}}{
An extract function for lrm objects from the \pkg{Design} or \pkg{rms} package.
}

\item{\code{extract.mer}}{
An extract function for mer objects from the (old) \pkg{lme4} package.
}

\item{\code{extract.negbin}}{
An extract function for negbin objects from the \pkg{MASS} package.
}

\item{\code{extract.nlmerMod}}{
An extract function for nlmerMod objects from the (new) \pkg{lme4} package.
}

\item{\code{extract.plm}}{
An extract function for plm objects from the \pkg{plm} package.
}

\item{\code{extract.pmg}}{
An extract function for pmg objects from the \pkg{plm} package.
}

\item{\code{extract.polr}}{
An extract function for polr objects from the \pkg{MASS} package.
}

\item{\code{extract.Relogit}}{
An extract function for Relogit objects from the \pkg{Zelig} package.
}

\item{\code{extract.rlm}}{
An extract function for rlm objects from the \pkg{MASS} package.
}

\item{\code{extract.rq}}{
An extract function for rq objects from the \pkg{quantreg} package.
}

\item{\code{extract.sclm}}{
An extract function for sclm objects from the \pkg{ordinal} package.
}

\item{\code{extract.simex}}{
An extract function for simex objects from the \pkg{simex} package.
}

\item{\code{extract.stergm}}{
An extract function for stergm objects from the \pkg{tergm} package.
}

\item{\code{extract.svyglm}}{
An extract function for svyglm objects from the \pkg{survey} package.
}

\item{\code{extract.systemfit}}{
An extract function for systemfit objects from the \pkg{systemfit} package.
}

\item{\code{extract.tobit}}{
An extract function for tobit objects from the \pkg{AER} package.
}
}}
\seealso{
\code{\link{texreg-package} \link{texreg} \link{extract}}
}
\keyword{methods}
