\name{.First.lib}
\alias{.First.lib}

\title{First Run}

\description{
  Loads native code and initializes .Random.seed.
}

\usage{
  .First.lib(libname, pkgname)
}

\arguments{
  \item{libname}{library}
  \item{pkgname}{section}
}

\value{
  a single uniform random number is returned invisibly.
}

\keyword{utilities}

