\name{get.presolveloops}
\alias{get.presolveloops}

\title{Get Presolve Loops}

\description{
Returns the number of times presolve must be done.
}

\usage{
get.presolveloops(lprec)
}

\arguments{
  \item{lprec}{an lpSolve linear programming model.}
}

\value{
a single integer value.
}

\references{url{http://lpsolve.sourceforge.net/5.5/index.htm}}

\author{Kjell Konis \email{kjell.konis@epfl.ch}}

\keyword{programming}

