% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_range.R
\name{get_coverage_in_range}
\alias{get_coverage_in_range}
\title{Get Coverage in Range}
\usage{
get_coverage_in_range(bwList, target_range, names, cvg_scaling = NULL)
}
\arguments{
\item{bwList}{vector: list of files to use to get coverage. Currently only accepts BigWig files}

\item{target_range}{GRange: specifies the range in which to get coverage}

\item{names}{vector: names to give each GRange coverage created}

\item{cvg_scaling}{vector: values by which to scale each coverage value. See get_cvg_scaling for more information.}
}
\value{
cvg_list GRangesList: coverage for each file supplied
}
\description{
Get Coverage in Range
}

