% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackview_functions.R
\name{get_exons}
\alias{get_exons}
\title{get exon information from IGIS}
\usage{
get_exons(target, src = c("symbol", "gene"), genome, org = c("human",
  "mouse"), ...)
}
\arguments{
\item{target}{character or GRange: gene symbol or GRange object for which to get exon information}

\item{src}{source of id, gene|symbol}

\item{genome}{}

\item{org}{organism, human or mouse}

\item{...}{additional argument passed to igisGetmart()}
}
\value{
a data frame of exon information
}
\description{
get exon information from IGIS
}
\author{
Jinfeng Liu
}

