% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_data_prep.R
\name{get_gene_range}
\alias{get_gene_range}
\title{get_gene_range}
\usage{
get_gene_range(exon.df)
}
\arguments{
\item{exon.df}{data.frame: the exons included in the annotation}
}
\value{
range.gene GRange: s
}
\description{
Make a GRange that cover the exon list
}
\author{
Justin Finkle
}

