% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{get_scaling_factor}
\alias{get_scaling_factor}
\title{Get Scaling Factor
Get the scaling factor for samples provided. This is assumed to be the number
of uniquely mapped reads}
\usage{
get_scaling_factor(file_info)
}
\arguments{
\item{file_info}{}
}
\description{
Get Scaling Factor
Get the scaling factor for samples provided. This is assumed to be the number
of uniquely mapped reads
}

