% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{get_subplot_ax_info}
\alias{get_subplot_ax_info}
\title{get_subplot_ax_info
Get the axes information in a plotly object. This is meant to be
called on an object created by the subplot function}
\usage{
get_subplot_ax_info(plotly_obj)
}
\arguments{
\item{plotly_obj}{}
}
\value{

}
\description{
get_subplot_ax_info
Get the axes information in a plotly object. This is meant to be
called on an object created by the subplot function
}

