% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{get_tx_annotation}
\alias{get_tx_annotation}
\title{Title}
\usage{
get_tx_annotation(db_object, range, tx_data, no_introns = FALSE)
}
\arguments{
\item{db_object}{}

\item{range}{}

\item{tx_data}{}

\item{no_introns}{}
}
\description{
Title
}

