% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_functions.R
\name{getChipFileInfo}
\alias{getChipFileInfo}
\title{getChipFileInfo}
\usage{
getChipFileInfo(samples_file, pairs_file)
}
\arguments{
\item{samples_file}{string}

\item{pairs_file}{string}

\item{sampl_name}{string}
}
\value{
list
}
\description{
.. content for the "Description" section of the documentation ..
}
\details{
.. content for "Details" section of the documentation ..
}
\author{
Justin Finkle
}

