% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackview_functions.R
\name{getCoverageInRange}
\alias{getCoverageInRange}
\title{Title}
\usage{
getCoverageInRange(bwList, target.range, names, scaling.factor = NULL)
}
\arguments{
\item{bwList}{}

\item{target.range}{}

\item{names}{}
}
\description{
Title
}

