% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggplot_RIGHT}
\alias{ggplot_RIGHT}
\title{Extract data object's name}
\usage{
ggplot_RIGHT(data, ...)
}
\arguments{
\item{data}{a data.frame object.}
}
\description{
Function to create data object name in ggplot function.
}
\examples{
\donttest{
obj <- RIGHT({ggplot(Theoph, aes(Time, conc, colour=Subject)) + geom_point()})
print(obj)
}
}
\seealso{
\code{\link{ggplot2}}
}

