\name{glmRob.misclass.w}
\alias{glmRob.misclass.w}
\alias{glmRob.misclass.F}
\alias{glmRob.misclass.f}
\alias{glmRob.misclass.G}
\alias{glmRob.misclass.g}

\title{Misclass Weights}

\description{
Compute the weights used by the consistant misclassification estimator.
}

\usage{
glmRob.misclass.w(u, lambda)
}

\arguments{
  \item{u}{a numeric vector.}
  \item{lambda}{a numeric value.}
}

\value{
a numeric vector.
}

\seealso{\code{\link{glmRob.misclass}}}

\keyword{robust}
\keyword{regression}


