% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plus_tracks_shiny.R
\name{heatmap_to_tracks_shiny}
\alias{heatmap_to_tracks_shiny}
\title{heatmap_to_tracks_shiny}
\usage{
heatmap_to_tracks_shiny(heatmap, track_function, link,
  title = "Heatmap linked to Genome Tracks", options = list(height = 1400))
}
\description{
heatmap_to_tracks_shiny
}
