\name{helpAdminister}
\alias{helpAdminister}
\docType{package}
\title{
MIfuns: Administrative Tasks
}
\description{Tools for administrative tasks.}

\details{
I want to ....
\itemize{
\item \link[=CLNR]{Delete} files and directories; log what was deleted to a text file
\item \link[=runlog]{Generate} individual run logs for NONMEM7 runs
\item \link[=rlog]{Generate} a combined run log from multiple NONMEM runs
\item \link[=accept]{Document} acceptance of an R installation
  }
}

