\name{helpAnalyze}
\alias{helpAnalyze}
\docType{package}
\title{
MIfuns: NONMEM Analysis
}
\description{Tools for preparing, executing, and analyzing NONMEM runs.}

\details{
I want to ....
\itemize{
\item \link[=NONR]{Run} NONMEM and create diagnostic plots
\item \link[=compute.cwres]{Compute} the conditional weighted residuals
\item \link[=rlog]{Generate} a combined run log from multiple NONMEM runs
\item \link[=metaSub.character]{Make} multiple, systematic substitutions in text files
\item \link[=as.keyed]{Create} and manipulate keyed data frames
\item \link[=nm]{Create} and manipulate nm objects
}
}

