\name{helpDataFrame}
\alias{helpDataFrame}
\docType{package}
\title{\pkg{MIfuns}: Working with Data Frames}
\description{Tools for processing data.frames.}
\details{
I want to ....
\itemize{
\item \link[=shuffle]{Move} columns in a data frame
\item \link[=fixedwidth]{Format} a data.frame as fixed-width
\item \link[=aug]{Add} named elements to a data.frame or list
}
}
