\name{helpModel}
\alias{helpModel}
\docType{package}
\title{MIfuns: Modeling}
\description{Tools for preparing, executing, and analyzing models.}

\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item \link[=helpPrepare]{Assemble} an analysis data set
\item \link[=metaSub.character]{Make} multiple, systematic substitutions in text files, e.g., control streams
\item \link[=NONR]{Run} NONMEM and create diagnostic plots
\item \link[=compute.cwres]{Compute} conditional weighted residuals in NONMEM 6
\item \link[=helpVisualize]{Visualize} model results
}
}

