\name{helpPharmacometric}
\alias{helpPharmacometric}
\docType{package}
\title{\link[=MIfuns]{MIfuns} \link[=helpPrepare]{Data}: Pharmacometrics}
\description{Tools for investigating pharmacometric properties.}
\details{
I want to ....
\itemize{
\item \link[=ibw]{Calculate} various clinical measurements, e.g., body size
\item \link[=css]{Calculate} one-compartment model properties
\item \link[=check.subjects]{Summarize} a table of subject data
\item \link[=Tmax]{Calculate} Tmax
\item \link[=Tmin]{Calculate} Tmin
\item \link[=AUC]{Calculate} AUC
}
}
