\name{helpPlot}
\alias{helpPlot}
\docType{package}
\title{
MIfuns: Plotting
}
\description{Tools for creating plots.}
\details{
I want to ....
\itemize{
\item \link[=NONR]{Run} NONMEM and create diagnostic plots
\item \link[=PLOTR]{Create} diagnostic plots for finished runs
\item \link[=covplot]{Make} \sQuote{forest plots} of covariate effects, etc.
\item \link[=plot.nm]{Plot} nm objects
\item \link[=xyplot.ext]{Plot} the parameter search history for a NONMEM7 run
}
}
