\name{helpPrepare}
\alias{helpPrepare}
\docType{package}
\title{metrumrg: Data Preparation}
\description{Tools for preparing pharmacometric analysis data sets.}
\details{
\link[=metrumrg]{metrumrg}

I want to ....
\itemize{
\item \link[=helpVector]{work with vectors}    
\item \link[=helpDataFrame]{work with data.frames}
\item \link[=helpMatrix]{work with matrices}
\item \link[=helpClasses]{use classes specialized for data assembly}
\item \link[=helpPharmacometric]{investigate pharmacometric properties}
\item \link[=helpStrategic]{perform strategic data manipulations}
\item \link[=merge.list]{merge recursively}
}
}
