\name{helpQuantify}
\alias{helpQuantify}
\docType{package}
\title{metrumrg: Statistics}
\description{Tools for calculating statistics.}
\details{
\link[=metrumrg]{metrumrg}

I want to ....
\itemize{
\item \link[=bin]{assign values in a vector to quantile bins}
\item \link[=snap]{coerce values to nearest of candidates}
\item \link[=bounds]{calculate normal quantiles at limits of a probability interval}
\item \link[=cv]{calculate coefficient of variation for normal or log-normal data}
\item \link[=percent]{calculate percent}
}
}
