\name{helpReport}
\alias{helpReport}
\docType{package}
\title{metrumrg: Reporting}
\description{Tools for summarizing and reporting pharmacometric analyses.}
\details{
\link[=metrumrg]{metrumrg}

I want to ....
\itemize{
\item \link[=params]{identify documented parameter names}
\item \link[=lookup]{lookup some other form of parameter names}
\item \link[=rlog]{generate a combined run log from multiple NONMEM runs}
\item \link[=partab]{create a model parameter table}
\item \link[=parameter2wiki]{convert parameter labels to wikimath}
\item \link[=wiki2latex]{convert wikimath to latex or plotmath}
\item \link[=justUnits]{manipulate wikimath elements}
\item \link[=iterations]{list the iteration statistics in the NONMEM output}
\item \link[=omegacor]{convert the NONMEM output omega or sigma covariance matrix to correlations}
\item \link[=cv]{calculate coefficients of variation for normal or log-normal data}
\item \link[=bounds]{calculate confidence intervals}
\item \link[=is.diagonal]{analyze NONMEM-style parameter labels}
\item \link[=ftable2data.frame]{reformat an ftable for printing with latex}
\item \link[=ltable.data.frame]{convert a data.frame for printing with latex}
\item \link[=as.pdf.data.frame]{render a data.frame to pdf format}
\item \link[=viewtex]{render file-based tex fragments as pdf for informal viewing}
\item \link[=xpath]{extract value of an xpath expression from an xml file}
}
}
