\name{helpSimulate}
\alias{helpSimulate}
\docType{package}
\title{
MIfuns: Simulation
}
\description{Tools for conducting simulations.}
\details{
I want to ....
\itemize{
\item \link[=NONR]{Run} NONMEM and create diagnostic plots
\item \link[=rlog]{Generate} a combined run log from multiple NONMEM runs
\item \link[=resample.data.frame]{Create} replicate datasets by stratified sampling with replacement    
\item \link[=metaSub.character]{Make} multiple, systematic substitutions in text files
\item \link[=simpar]{Create} parameters for simulation from uncertainity
\item \link[=as.matrix.halfmatrix]{Convert} a vectorized triangular matrix to a symmetric matrix
}
}
