\name{helpVisualize}
\alias{helpVisualize}
\docType{package}
\title{MIfuns: Plotting}
\description{Tools for creating plots.}
\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item \link[=PLOTR]{Create} diagnostic plots for finished model runs
\item \link[=covplot]{Make} \sQuote{forest plots} of covariate effects
\item \link[=panel.stratify]{Handle} each level of a stripplot independently 
\item \link[=plot.nm]{Plot} nm objects
\item \link[=xyplot.ext]{Plot} the parameter search history for a NONMEM7 run
}
}
