\name{holzinger}
\alias{holzinger}
\title{Description of holzinger data}
\description{
Description of data sets.
}

\usage{
data(holzinger)

}

 \arguments{
   \item{holzinger}{100 simulated cases from the infamous Holzinger correlation matrix}
 }

\author{Phil Chalmers \email{rphilip.chalmers@gmail.com}}

\keyword{data} 