\name{hugin.domain}
\alias{hugin.domain}

\title{Create an RHugin Domain Object}

\description{
Create an RHugin domain object.
}

\usage{
hugin.domain()
}

\value{
an RHugin domain object.
}

\author{Kjell Konis \email{kjell.konis@epfl.ch}}

\seealso{
  \code{\link{add.node}},
  \code{\link{add.edge}}.
}

\examples{
# Create an empty RHugin domain
AppleTree <- hugin.domain()
}

\keyword{programming}


