% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackview_functions.R
\name{igisExonlist}
\alias{igisExonlist}
\title{get exon information from IGIS}
\usage{
igisExonlist(target, src = c("symbol", "gene"), genome, org = c("human",
  "mouse"), ...)
}
\arguments{
\item{src}{source of id, gene|symbol}

\item{org}{organism, human or mouse}

\item{...}{additional argument passed to igisGetmart()}

\item{ids}{vector of ids to retrive the exon information}
}
\value{
a data frame of exon information
}
\description{
get exon information from IGIS
}
\author{
Jinfeng Liu
}

