\name{.twsIncomingMSG}
\alias{.twsIncomingMSG}
\alias{.twsOutgoingMSG}
\alias{.twsERR}
\alias{.lastRequest}
\alias{reqExecutions}
\alias{twsExecutionFilter}
\alias{twsDEBUG}
\title{ Internal TWS-API MSG and ERR List }
\description{
Internal List of MSG Codes and Undocumented (Experimental) Functions
}
\keyword{ utilities }
