\name{is.quantmod}
\alias{is.quantmod}
\alias{is.quantmodResults}
\title{ Test If Object of Type quantmod }
\description{
Test if object is of type \code{quantmod} or
\code{quantmodResults}.
}
\usage{
is.quantmod(x)
is.quantmodResults(x)
}
\arguments{
  \item{x}{ object to test }
}
\value{
Boolean TRUE or FALSE
}
\author{ Jeffrey A. Ryan }
\seealso{ \code{\link{specifyModel}}, \code{\link{tradeModel}} }
\keyword{ utilities }
