\name{llratio}\docType{method}\alias{llratio}\alias{llratio,depmix.fitted-method}\alias{likelihoodratio}\alias{show,llratio-method}\title{Log likelihood ratio test on two fitted models}\description{Performs a log likelihood ratio test on two fitted\code{depmix} models.}\usage{		llratio(basemodel,constrainedmodel,...)	}\arguments{ 		\item{basemodel}{Fitted depmix model.}		\item{constrainedmodel}{Fitted depmix model.}		\item{...}{Not currently used.}	}\value{	\code{llratio} returns an object of class \code{llratio} which has slots: 		\item{value}: Minus twice the loglikelihood difference. 		\item{df}: The degrees of freedom, ie the difference in number of freely	estimated paraemters between the models. 		The print method shows the value, 	}\author{Ingmar Visser}\keyword{methods}