\name{lmRob.ga}
\alias{lmRob.ga}
\title{
  Estimation of S-Estimates with Genetic Algorithm 
}
\description{
Computes the robust S-estimates using a genetic algorithm. It is used by  
\code{lmRob}, not supposed to be called by the users directly. 
}

\keyword{robust}

% docclass is function
% Converted by Sd2Rd version 37351.
