% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data.R
\name{load_snp_data}
\alias{load_snp_data}
\title{Load SNP data}
\usage{
load_snp_data(file.name, sep = "\\t")
}
\arguments{
\item{file.name}{str: the file containig the snp information}

\item{sep}{}
}
\value{
snp.gr GRanges:
}
\description{
Function to load SNP information from GWAS catalog v1.0
}
