\name{lsfTmpDir}
\alias{lsfTmpDir}
\title{Rlsf working directory}
\description{
Create a working directory for Rlsf
}
\usage{
odfTmpDir()
}
\details{
This function determines the temp directory and creates a subdirectory. If a subdirectory cannot be created, an error is thrown.
}
\value{
a character string containing the path
}
\author{Max Kuhn}

\keyword{utilities}

