% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_ly.R
\name{make_browserly_function}
\alias{make_browserly_function}
\title{Title}
\usage{
make_browserly_function(cvg_files, tx_data, sample_names = names(cvg_files),
  hm_thresh = 4, type = NULL, cvg_scaling = rep(1, length(cvg_files)))
}
\arguments{
\item{cvg_files}{}

\item{tx_data}{}

\item{hm_thresh}{}

\item{type}{}

\item{cvg_scaling}{}

\item{genome}{}
}
\description{
Title
}

