% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gviz_wrappers.R
\name{make_data_tracks}
\alias{make_data_tracks}
\title{Make Data Tracks}
\usage{
make_data_tracks(cvg.L, gr, genome, chr, ymax, sync = FALSE,
  bg.title = "black", scale.group = 1, colors = NULL, type = NULL,
  binwidth = 1000, val = "score", scaling = NULL, kwargs = list())
}
\arguments{
\item{cvg.L}{GRangesList containing coverage score}

\item{gr}{GRange to plot}

\item{genome}{str}

\item{chr}{str}

\item{bg.title}{str: color for track panel backgrounds}

\item{colors}{list-optional: colors for coverage tracks}

\item{type}{str: 'hist' for multiple coverage tracks, 'heatmap' for condensed view}

\item{binwidth}{scalar}

\item{val}{str: value in cvg.L elements to use for binning coverage. Default is "score"}

\item{scaling}{}

\item{score.max}{scalar: scaling}

\item{dtrack.kwargs}{list-optional: additional arguments for DataTrack. These should only be used for global arguments that will apply to all datatracks}
}
\value{
dtrack list: DataTrack objects
}
\description{
Builds Gviz DataTrack objects for plotting based on coverage
}
\author{
Justin Finkle
}

