% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_ly.R
\name{make_subplots}
\alias{make_subplots}
\title{make_subplots
Make subplots of coverage tracks}
\usage{
make_subplots(grl, type, ...)
}
\arguments{
\item{grl}{GRangesList: a named list. Each GRanges object in the list will be plotted on a separate axis}

\item{type}{str: type of plot to use}

\item{...}{additional parameters to pass to browserly_cvg_track}
}
\value{

}
\description{
make_subplots
Make subplots of coverage tracks
}

