% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackview_functions.R
\name{make_track_function}
\alias{make_track_function}
\title{Title}
\usage{
make_track_function(range, genome, org, cvg_files, scaling_factor, symbol,
  transcripts.gr, ymax, bg.title = "black", colors = NULL, sync = FALSE,
  type = NULL, scale.group = 1, hm.thresh = 4, scaling = NULL,
  hm.binsize = 1000, snp.gr = NULL, tss.gr = NULL, ucsc = FALSE,
  gatrack.kwargs = list(), grtrack.kwargs = list(),
  tsstrack.kwargs = list(), snptrack.kwargs = list(),
  dtrack.kwargs = list(), plot.kwargs = list())
}
\arguments{
\item{range}{}

\item{genome}{}

\item{org}{}

\item{cvg_files}{}

\item{scaling_factor}{}

\item{symbol}{}

\item{transcripts.gr}{}

\item{ymax}{}

\item{bg.title}{}

\item{colors}{}

\item{sync}{}

\item{type}{}

\item{scale.group}{}

\item{hm.thresh}{}

\item{scaling}{}

\item{hm.binsize}{}

\item{snp.gr}{}

\item{tss.gr}{}

\item{ucsc}{}

\item{gatrack.kwargs}{}

\item{grtrack.kwargs}{}

\item{tsstrack.kwargs}{}

\item{snptrack.kwargs}{}

\item{dtrack.kwargs}{}

\item{plot.kwargs}{}
}
\description{
Title
}

