% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{match_tx_to_gene}
\alias{match_tx_to_gene}
\title{Title}
\usage{
match_tx_to_gene(txdb, org_db, str_regex = "[[:digit:]]+")
}
\arguments{
\item{txdb}{}

\item{org_db}{}

\item{str_regex}{}
}
\description{
Title
}
