% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalOrderMeasureFunctions.R
\name{measurePolarization}
\alias{measurePolarization}
\title{Polarization collective order}
\usage{
measurePolarization(frameData)
}
\arguments{
\item{frameData}{A data frame of 4 columns (x, y, u, v) specifying the position and velocities items in group.}
}
\value{
A scalar between -1 and 1 specifying the collective order.
}
\description{
Polarization collective order
}
