% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalOrderMeasureFunctions.R
\name{measureRotation}
\alias{measureRotation}
\title{Rotation collective order}
\usage{
measureRotation(frameData)
}
\arguments{
\item{frameData}{A data frame of 4 columns (x, y, u, v) specifying the position and velocities items in group.}
}
\description{
Rotation collective order
}
