\name{metrumrg-package}
\alias{metrumrg-package}
\alias{metrumrg}
\docType{package}
\title{Pharmacometric Tools for Data Preparation, Modeling, Simulation, and Reporting}
\description{
This package includes pharmacometric tools for common data preparation tasks, stratified
bootstrap resampling of data sets, NONMEM control stream
creation/editing, NONMEM model execution, creation of standard and
user-defined diagnostic plots, execution and summary of bootstrap and
predictive check results, implementation of simulations from posterior
parameter distributions, reporting of output tables, and creation of
detailed analysis logs.
}
\details{

I want to ....
\itemize{
\item \link[=helpPrepare]{prepare} or manipulate data
\item \link[=helpModel]{model} data in NONMEM
\item \link[=helpSimulate]{simulate} from existing models
\item \link[=helpVisualize]{plot} model inputs or outputs
\item \link[=helpQuantify]{calculate} general statistics
\item \link[=helpReport]{report} modeling and simulation results systematically
\item \link[=helpAdminister]{administer} related files and file systems
\item \link[=helpList]{list} all functions
\item \link[=helpBrowse]{browse} extended examples
}  
\tabular{ll}{
Package: \tab metrumrg\cr
Type: \tab Package\cr
Version: \tab 5.19\cr
Date: \tab 2011-05-30\cr
License: \tab GPL\cr
}
}
  
  \author{Tim Bergsma, Bill Knebel, Leonid Gibiansky, Natalie Hsiang. Maintainer: Tim Bergsma <timb@metrumrg.com>}
  \references{\url{http://metrumrg.googlecode.com}}
  \keyword{manip}
