\name{MIfuns-package}
\alias{MIfuns-package}
\alias{MIfuns}
\docType{package}
\title{Pharmacometric Tools for Data Preparation, Modeling, Simulation, and Reporting}
\description{
This package includes pharmacometric tools for common data preparation tasks, stratified
bootstrap resampling of data sets, NONMEM control stream
creation/editing, NONMEM model execution, creation of standard and
user-defined diagnostic plots, execution and summary of bootstrap and
predictive check results, implementation of simulations from posterior
parameter distributions, reporting of output tables, and creation of
detailed analysis logs.
}
\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item \link[=helpPrepare]{prepare or manipulate data}
\item \link[=helpModel]{model data in NONMEM}
\item \link[=helpSimulate]{conduct simulations}
\item \link[=helpVisualize]{create plots of model inputs or outputs}
\item \link[=helpQuantify]{calculate statistics}
\item \link[=helpReport]{generate report material}
\item \link[=helpAdminister]{perform related filesystem tasks}
\item \link[=helpList]{see a list of available functions}
}  
\tabular{ll}{
Package: \tab MIfuns\cr
Type: \tab Package\cr
Version: \tab 4.0.5\cr
Date: \tab 2010-06-09\cr
License: \tab GPL\cr
}
}
  
  \author{Developed by Metrum Institute (Bill Knebel, Tim Bergsma, Leonid Gibianski, others). Maintainer: Tim Bergsma <timb@metrumrg.com>}
  
  \references{http://mifuns.googlecode.com}
  \keyword{manip}
