% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_single_coverage_heatmap.R
\name{multi_coverage_heatmap}
\alias{multi_coverage_heatmap}
\title{multi_coverage_heatmap}
\usage{
multi_coverage_heatmap(mats, x = iHeatmapR:::default_x(mats[[1]]),
  y = iHeatmapR:::default_y(mats[[1]]), row_order = c("none", "hclust",
  "kmeans", "groups", "signal"), k = NULL, groups = NULL,
  clust_dist = stats::dist, clust_include = 1, signal = NULL,
  name = "Signal", summary = TRUE, source = "HM", scale = c("rows",
  "none"), scale_method = c("normalize", "standardize", "center"),
  share_z = TRUE, ...)
}
\arguments{
\item{x}{x axis labels}

\item{y}{y axis labels}

\item{row_order}{row order method}

\item{k}{k to use for kmeans clustering or cutting heirarchical clustering}

\item{groups}{pre-determined groups for rows}

\item{signal}{signal along row}

\item{name}{name of colorbar}

\item{summary}{make summary plot, boolean}

\item{source}{source name in plotly}

\item{scale}{scale rows? or none}

\item{scale_method}{method to use for scaling}

\item{share_z}{share colorbar between heatmaps}

\item{mat}{coverage matrix}
}
\value{
iHeatmap object
}
\description{
multi_coverage_heatmap
}
\author{
Alicia Schep
}

