% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_coverage_shiny.R
\name{multi_coverage_plot}
\alias{multi_coverage_plot}
\title{mulit_coverage_plot}
\usage{
multi_coverage_plot(inputs, regions, region_names, binsize = 50, up = 2000,
  down = 2000, col_aggs = NULL, row_aggs = NULL, groups = NULL,
  cvg_files = NULL, aggregate_normalization = c("localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar", "none"),
  heatmap_normalization = c("localRms", "localMean", "localNonZeroMean",
  "PercentileMax", "scalar", "none"), genome = "GRCm38", org = "mouse", ...)
}
\description{
mulit_coverage_plot
}
\details{
More details will go here.
}
\author{
Alicia Schep
}

