% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_single_coverage_heatmap.R
\name{new_single_coverage_heatmap}
\alias{new_single_coverage_heatmap}
\title{new_single_coverage_heatmap}
\usage{
new_single_coverage_heatmap(mat, x = iHeatmapR:::default_x(mat),
  y = iHeatmapR:::default_y(mat), row_order = c("none", "hclust", "kmeans",
  "groups", "signal"), k = NULL, groups = NULL, clust_dist = stats::dist,
  signal = NULL, name = "Signal", summary = TRUE, source = "HM",
  scale = c("rows", "none"), scale_method = c("normalize", "standardize",
  "center"), ...)
}
\arguments{
\item{mat}{coverage matrix}

\item{x}{x axis labels}

\item{y}{y axis labels}

\item{row_order}{row order method}

\item{k}{k to use for kmeans clustering or cutting heirarchical clustering}

\item{groups}{pre-determined groups for rows}

\item{signal}{signal along row}

\item{name}{name of colorbar}

\item{summary}{make summary plot, boolean}

\item{source}{source name in plotly}
}
\value{
list with three components, 1) plot -- function to plot
2) row_order -- row order used, 3) dendro hclust object if heirarchical clustering
performed
}
\description{
new_single_coverage_heatmap
}
\author{
Alicia Schep
}

