\name{odecision-class}
\Rdversion{1.1}
\docType{class}
\alias{odecision-class}
\alias{odecision}

\title{Class \code{"odecision"}}
\description{
Definition container for \code{\link{CheckDecisions}}
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("odecision", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{dectable}:}{Object of class \code{"data.frame"} ~~ }
    \item{\code{condition}:}{Object of class \code{"list"} ~~ }
    \item{\code{effect}:}{Object of class \code{"list"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "odecision" in the signature.
}
\references{
%%  ~~put references to the literature/web site here~~
}
\author{
T. Rintala \email{teemu.rintala.a@gmail.com}
}
\note{
See also:
\url{http://en.opasnet.org/}
}

\seealso{
%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("odecision")
}
\keyword{classes}
