\name{output}
\alias{output}

\title{
Generating Basic Summary Statistics In Multidimensions
}

\description{
Producing basic summary statistics of mean, standard deviation and the first five modes from a sample, which is drawn via the Fu-Wang algorithm and the Wang-Lee algorithm, in each dimension of a target function. 
}

\usage{
output(x)
}

\arguments{
\item{x}{
a \code{data.frame}, contains a sample which is drawn via either the Fu-Wang algorithm or the Wang-Lee algorithm 
}
}

\author{
Liqun Wang and Chel Hee Lee Lee
}

