\name{pan-package}
\alias{pan-package}
\alias{pan}
\docType{package}
\title{
PAN: Phylogenetic Analysis
}
\description{
Tools to perform phylogenetic analyses.
}
\details{
\tabular{ll}{
Package: \tab pan\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-04-16\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
It's going to do a bunch of stuff but for now it just returns the number of transitions and transversions in an alignment.
}
\author{
Francois Michonneau after recycling and modifying code written by Emmanuel Paradis.

Maintainer: Francois Michonneau  <francois.michonneau@gmail>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
