    \name{peakbound}
          \Rdversion{1.0}
          \alias{basecountimport}
          \title{Return Exact Peak Bounds Given numeric string of Basecounts 'x'}
          \description{
            Imports basecounts from significant windows using generated coordinate file from getsigwindows(), each string of basecounts much be length 1500 (for now)
          }
          \usage{basecountimport(inputfile, coordfile, outputfile, chromosome='all')}
          \arguments{
            \item{x}{string of numeric basecounts}
          }
          \seealso{
            \code{\link{save}}.
          }
          \examples{

   
          }
          \keyword{file}

