% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_ly.R
\name{plot_browserly_tracks}
\alias{plot_browserly_tracks}
\title{Title}
\usage{
plot_browserly_tracks(target_range, tx_data, cvg, hm_thresh = 4,
  type = NULL)
}
\arguments{
\item{tx_data}{}

\item{hm_thresh}{}

\item{type}{}

\item{txdb}{}

\item{range}{}
}
\description{
Title
}

