% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly.R
\name{plot_single_locus}
\alias{plot_single_locus}
\title{Title}
\usage{
plot_single_locus(target_range, tx_data, cvg, hm_thresh = 4,
  stacking = "dense", sync_y = TRUE, type = NULL, xlabel = NULL,
  snps = NULL, ...)
}
\arguments{
\item{target_range}{}

\item{tx_data}{}

\item{hm_thresh}{}

\item{stacking}{}

\item{sync_y}{}

\item{type}{}

\item{...}{}
}
\description{
Title
}

