% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackview_functions.R
\name{plot_track_view}
\alias{plot_track_view}
\title{plot genomic coverage along the gene
Adapted from gChipseq function of the same name. Credit to Jinfeng Liu
plot genomic coverage along the gene}
\usage{
plot_track_view(grl, dat.exon, target.range, genome, ymax, symbol,
  bg.title = "grey50", colors = NULL, sync = FALSE, type = NULL,
  scale.group = 1, hm.thresh = 4, scaling = NULL, hm.binsize = 1000,
  showSNPs = FALSE, snp.gr = NULL, tss.gr = NULL,
  dtrack.kwargs = list(), gtrack.kwargs = list(), grtrack.kwargs = list(),
  snptrack.kwargs = list(), ...)
}
\arguments{
\item{grl}{GRangesList: typically the results of importing wig/bw/... files for each sample}

\item{dat.exon}{data.frame: exon information}

\item{target.range}{GRange: range to plot}

\item{genome}{str: genome build, e.g. hg19, GRCh38, etc.}

\item{ymax}{vector-optional: ymax for the data tracks}

\item{symbol}{str-optional: symbol to plot}

\item{colors}{str-optional: colors for coverage data tracks}

\item{sync}{logical-optional: whether to sync the ymax for all data tracks}

\item{type}{str-optional: type of datatrack to plot. Currently only 'hist' and 'heatmap' are supported}

\item{scale.group}{scalar-optional: groups to use the same y scale}

\item{hm.thresh}{integer-optional: threshold above which a heatmap is plotted. Can change threshold or override with type}

\item{dtrack.kwargs}{list-optional: additional arguments for DataTrack. These should only be used for global arguments that will apply to all datatracks}

\item{gtrack.kwargs}{list-optional: additional arguments for GenomeAxisTrack}

\item{grtrack.kwargs}{list-optional: additional arguments for GeneRegionTrack}

\item{snptrack.kwargs}{list-optional: additional arguments for the SNP DataTrack}

\item{bg.title:}{str-optional: background color for the title panel}
}
\value{
ptracks list: track objects plotted by Gviz
}
\description{
plot genomic coverage along the gene
Adapted from gChipseq function of the same name. Credit to Jinfeng Liu
plot genomic coverage along the gene
}
\author{
Justin Finkle
}

