\name{plotTTree}
\alias{plotTTree}
\title{Plot a transmission tree}
\usage{
plotTTree(ttree, w.shape = 2, w.scale = 1)
}
\arguments{
  \item{ttree}{Transmission tree}

  \item{w.shape}{Shape parameter of the Gamma probability
  density function representing the generation length w}

  \item{w.scape}{Scale parameter of the Gamma probability
  density function representing the generation length w}
}
\description{
Plot a transmission tree
}

