\name{print.summary.RHuginDomain}
\alias{print.summary.RHuginDomain}

\title{summary.RHuginDomain Print Method}

\description{
  Print an RHuginDomain summary object.
}

\usage{
\method{print}{summary.RHuginDomain}(x, ...)
}

\arguments{
  \item{x}{an RHuginDomain summary object.}
  \item{\dots}{additional arguments are ignored.}
}

\value{
  \code{x} is invisibly returned.
}

\author{Kjell Konis \email{kjell.konis@icloud.com}}

\keyword{methods}


