\name{probTTree}
\alias{probTTree}
\title{Calculates the log-probability of a transmission tree
(assumes w is Gamma(2,1))}
\usage{
probTTree(ttree, R)
}
\arguments{
  \item{ttree}{Transmission tree}

  \item{R}{Basic reproduction number}
}
\value{
Probability of the transmission tree
}
\description{
Calculates the log-probability of a transmission tree
(assumes w is Gamma(2,1))
}

