% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R2rflow.R
\name{R2rflow}
\alias{R2rflow}
\title{Convert R code to rflow file}
\usage{
R2rflow(text = NULL, file = NULL, output = NULL, expand = TRUE)
}
\arguments{
\item{text}{A text containing R code.}

\item{file}{A file containing R code.}

\item{output}{An output rflow file. If a 'file' argument is specified, 'output' argument is
assumed to have the same file name as in 'file' but with rflow extension.}

\item{expand}{Should diagram be expanded up to a single command?}
}
\value{
xml formated R code
}
\description{
Takes a file or text argument and prints out rflow xml
}
\author{
Lukasz Daniel
}

