\name{read.HTSeqCounts}
     \alias{read.HTSeqCounts}
     \title{Read counts output from HTSeq script.}
     \description{
       This function reads the output files from HTSeq python script \code{X.py}, it provides an ExonCountSet object.
     }
     \usage{
     read.HTSeqCount(countfiles, design, aggregatefile=NULL)
     }
     \arguments{
       \item{countfiles}{A string vector containing the output files with the paths from \code{aggregate_genes.py} to read.}
       \item{design}{A vector of strings or factors with the conditions foreach countfile.}
       \item{aggregatefile}{An aggregatefile generated by \code{count.py} necessary for the visualization of the data but not obligatory to test for alternative exon usage.}
     }
     \value{
	An ExonCount object.
     }
     \examples{
	countfiles <- dir(system.file("files/", package="DEXSeq"), pattern="counts")
	aggregatefile <- dir(system.file("files/", package="DEXSeq"), pattern="aggregate")
	aggregatefile <- paste(system.file("files/", package="DEXSeq"), aggregatefile, sep="")
	countfiles <- paste(system.file("files/", package="DEXSeq"), countfiles, sep="")
	ecs <- read.HTSeqCounts(c("untreated3.counts", "untreated4.counts", "gsm461180.counts", "gsm461181.counts"), c("un", "un", "gs", "gs"), aggregatefile="aggregates_dm3_refgene2.gff")
     }
\keyword{ExonCountSet}
