% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChipseq_functions.R
\name{readMacsPeaks}
\alias{readMacsPeaks}
\title{Read peaks from MACS output}
\usage{
readMacsPeaks(file, fc, minTags, neglog10pThreshold, fdr, use.summit = FALSE,
  as.Granges = TRUE, broad = FALSE)
}
\arguments{
\item{file}{filename}

\item{fc}{Fold-enrichment threshold}

\item{minTags}{Minimum number of tags in the peak}

\item{neglog10pThreshold}{Threshold value for the "-log10(p.value)" reported in MACS output}

\item{fdr}{FDR cut-off}

\item{use.summit}{use summit position instead of peak start/end position as main positions}

\item{as.Granges}{Logical. If the peaks should be returned as a GRanges object.}

\item{broad}{whether the input is macs2's broadpeak file}
}
\value{
data frame of macs peaks, with cleaned up column names and FDR values converted from percentage to relative value (between 0 and 1)
}
\description{
Read peaks from MACS output, both v1.4 and v2. Column names in v1 and v2 will be harmonized, use the peak start/end position as the main positions by default
}
\author{
Suchit Jhunjhunwala, Jinfeng Liu
}
