% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gChipseq_functions.R
\name{renameMacsColname}
\alias{renameMacsColname}
\title{rename MACS output column names}
\usage{
renameMacsColname(x)
}
\arguments{
\item{x}{a vector of MACS column names}
}
\value{
a vector of fixed column names
}
\description{
rename MACS output column names
}
\author{
Jinfeng Liu
}
