\name{ReutNews}
\docType{data}
\alias{ReutNews}
\title{Five exemplary news articles from the Reuters-21578 XML data set}
\description{
  This data set holds five news articles with additional meta
  information from the Reuters-21578 XML data set.
}
\usage{data(ReutNews)}
\format{A list with five text documents.}
\source{Reuters-21578 Text Categorization Collection Distribution 1.0
  (XML format)}
\references{
  Lewis, David (1997) \emph{Reuters-21578 Text Categorization
    Collection Distribution 1.0}.
  \url{http://kdd.ics.uci.edu/databases/reuters21578/reuters21578.html}
}
\keyword{datasets}
