\name{RHExample}
\alias{RHExample}

\title{Run RHugin Example (demo) Scripts}

\description{
This function runs the demo scripts in the RHugin package silently.  It is used to build networks for the examples in the help files.
}

\usage{
RHExample(topic)
}

\arguments{
  \item{topic}{a character string specifying the name of an RHugin demo script.}
}

\value{
This function is used for it's side effect: building a network for use in an example.
}

\author{Kjell Konis \email{kjell.konis@epfl.ch}}

\keyword{utilities}


