% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_heatmap.R
\name{rna_heatmap}
\alias{rna_heatmap}
\title{rna_heatmap}
\usage{
rna_heatmap(mat, x = iHeatmapR:::default_x(mat),
  y = iHeatmapR:::default_y(mat), row_order = c("hclust", "none", "kmeans",
  "groups"), col_order = c("groups", "none", "hclust", "kmeans"),
  row_groups = NULL, col_groups = NULL, row_k = NULL, col_k = NULL,
  row_clust_dist = stats::dist, col_clust_dist = stats::dist,
  name = "RNA-seq", source = "HM", scale = c("rows", "cols", "none"),
  scale_method = c("standardize", "center", "normalize"), x_labels = x,
  y_labels = NULL, x_title = NULL, y_title = NULL, ...)
}
\arguments{
\item{mat}{matrix of values to be plotted as heatmap}

\item{x}{x xaxis labels, by default colnames of mat}

\item{y}{y axis labels, by default rownames of mat}

\item{row_order}{how to order rows?  See Details}

\item{col_order}{how to order columns? See Details}

\item{row_groups}{groups for rows, overriden if row_k is set and row_order is kmeans or hclust}

\item{col_groups}{groups for columns, overriden if col_k is set and col_order is kmeans or hclust}

\item{row_k}{number of clusters for rows, needed if row_order is kmeans or optional if hclust}

\item{col_k}{number of clusters for columns, needed if row_order is kmeans or optional if hclust}

\item{row_clust_dist}{distance function to use for row clustering if hierarchical clustering}

\item{col_clust_dist}{distance function to use for column clustering if hierarchical clustering}

\item{name}{Name for colorbar}

\item{source}{source name, useful for shiny}

\item{scale}{scale matrix by rows, cols or none}

\item{scale_method}{what method to use for scaling, either standardize, center, normalize}

\item{x_labels}{axis labels for x axis (default is x)}

\item{y_labels}{axis labels for y axis (default is NULL)}

\item{x_title}{x axis title (default is NULL)}

\item{y_title}{y axis title (default is NULL)}

\item{...}{additional argument to make_main_hm}
}
\value{
iHeatmap object, which can be printed or passed to \code{\link{plot_iHeatmap}} to
generate an interactive graphic
}
\description{
rna_heatmap
}
\author{
Alicia Schep
}

