% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{run_full_counterfactual}
\alias{run_full_counterfactual}
\title{run_full_counterfactual}
\usage{
run_full_counterfactual(filename, init, inter, trans, beta_type,
  susceptible_type, beta_pars, susceptible_pars, ntime, ntrial = 1000)
}
