\name{SanMartinoPPts}
\docType{data}
\alias{SanMartinoPPts}
\title{
San Martino, ts of daily precipitation.
}
\description{
Daily time series of precipitation, at station San Martino di Castrozza, Trento Province, Italy, with data from 01/Jan/1921 to 31/Dec/1990.
}
\usage{
data(SanMartinoPPts)
}

%%\details{
%%
%%}
\format{
zoo object.
}
\source{
Provided by MeteoTrentino, Trento, Italy (via prof. Riccardo Rigon). \cr
These data are intended to be used for research purposes only, being distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY.
}

%%\references{
%% ~put references to the literature/web site here ~
%%}
\keyword{datasets}
