% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_coverage_shiny.R
\name{single_coverage_plot}
\alias{single_coverage_plot}
\title{single_coverage_plot}
\usage{
single_coverage_plot(input, regions, region_names, binsize = 50, up = 2000,
  down = 2000, row_agg = NULL, groups = NULL, cvg_files = NULL,
  heatmap_normalization = c("localRms", "localMean", "localNonZeroMean",
  "PercentileMax", "scalar", "none"), genome = "GRCm38", org = "mouse", ...)
}
\arguments{
\item{input}{either coverage matrix or input}
}
\value{
shiny app
}
\description{
single_coverage_plot
}
\author{
Alicia Schep
}

