\name{sqldf-package}
\alias{sqldf-package}
\docType{package}
\title{
sqldf package overview
}
\description{
Provides an easy way to perform SQL selects on R data frames.  
}
\details{
The package contains a single function \link{sqldf} whose help file
contains more information and examples.
}
\references{ 
The \link{sqldf} help page contains the primary documentation.
The sqldf home page \url{http://code.google.com/p/batchfiles/} contains
links to SQLite pages that may be helpful in formulating queries.
}
\keyword{ package }


