\name{star}
\alias{star}

\title{Replace Asterisk in x With y}
\description{
 helper function to dereference '*'.
}
\usage{
star(x, y)
}

\arguments{
  \item{x}{character}
  \item{y}{character}
}
\value{character}
\references{mifuns.googlecode.com}
\author{Tim Bergsma}
\seealso{\code{\link{PLOTR}}}
\keyword{manip}

