\name{tilePlot}
\alias{tilePlot}
\title{Add small plot on existing plot.}
\usage{
  tilePlot(expr, xlim, ylim, pol)
}
\arguments{
  \item{expr}{Expression of plot commands producing new
  plot.}

  \item{xlim}{New plot x-range.}

  \item{ylim}{New plot y-range.}

  \item{pol}{Polygon defining the sub-window wrt. large
  plot coordinate system.}
}
\value{
  Output of evaluated \code{expr}
}
\description{
  Add new plot on existing plot.
}
\details{
  Add new plot on existing plot.
}

