\name{ttreeFromFullTree}
\alias{ttreeFromFullTree}
\title{Extracts transmission tree from a combined phylogenetic/transmission tree}
\usage{
  ttreeFromFullTree(fulltree)
}
\arguments{
  \item{fulltree}{Combined tree}
}
\value{
  transmission tree
}
\description{
  Extracts transmission tree from a combined
  phylogenetic/transmission tree
}

