\name{undocumented}
\alias{.First.lib}

\title{Undocumented Functions}

\description{
These lpSolve functions are provided without documentation. 
}

\usage{
.First.lib(libname, pkgname)
}

\arguments{
  \item{libname}{libname}
  \item{pkgname}{pkgname}
}

\value{
usually something.
}

\references{url{http://lpsolve.sourceforge.net/5.5/index.htm}}

\author{Kjell Konis \email{kjell.konis@epfl.ch}}

\keyword{programming}

