\name{Viewer-class}
\Rdversion{1.1}
\docType{class}
\alias{Viewer-class}

\title{Class "Viewer" ~~~ }
\description{
  A class representing a current view of a data source.
  This is the View part of the MVC components of the data viewer.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "Viewer" in the signature.
}
\author{
  Paul Murrell
}
\seealso{
	~~objects to See Also as \code{\link{~~fun~~}}, ~~~
	or \code{\linkS4class{CLASSNAME}} for links to other classes
}
\examples{
showClass("Viewer")
}
\keyword{classes}
