% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_single_coverage_heatmap.R
\name{add_coverage_heatmap}
\alias{add_coverage_heatmap}
\title{add_coverage_heatmap}
\usage{
add_coverage_heatmap(p, mat, x = iHeatmapR:::default_x(mat),
  groups = p$row_groups, signal = rowMeans(mat, na.rm = TRUE),
  plot_signal = TRUE, name = "Coverage", signal_name = "Avg. Coverage",
  summary = TRUE, scale_method = c("localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar", "none"), pct = 0.95,
  scale_factor = 1, ticktext = TRUE, start = x[1], end = default_end(x),
  xlab = "Position", font = list(size = 8), ...)
}
\arguments{
\item{mat}{coverage matrix}

\item{x}{x axis labels}

\item{groups}{pre-determined groups for rows}

\item{signal}{signal along row}

\item{name}{name of colorbar}

\item{summary}{make summary plot, boolean}

\item{scale_method}{"normalize", "standardize", "center"}

\item{scale}{"rows" or "none"}
}
\value{
iHeatmap object
}
\description{
add_coverage_heatmap
}
\author{
Alicia Schep
}

