% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly.R
\name{add_tx_shapes}
\alias{add_tx_shapes}
\title{Title}
\usage{
add_tx_shapes(plotly_obj, tx_info, grt_ax, target_range, y_scaling = 0)
}
\arguments{
\item{plotly_obj}{}

\item{tx_info}{}

\item{grt_ax}{}

\item{target_range}{}
}
\description{
Title
}

