% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_profile.R
\name{aggregate_profile_plot}
\alias{aggregate_profile_plot}
\title{aggregate_profile_plot}
\usage{
aggregate_profile_plot(cvg_mats, positions = colnames(cvg_mats[[1]]),
  ylab = "ChIP Signal", xlab = "Position", showlegend = TRUE)
}
\description{
aggregate_profile_plot
}
\author{
Alicia Schep
}

