\name{anova.ALKmodel}
\alias{anova.ALKmodel}
\title{Compute approximate likelihood ratio test for the equality of two age-length keys.}
\usage{
  anova.ALKmodel(m1, m2)
}
\arguments{
  \item{m1}{An object of class 'ALKmodel'}

  \item{m2}{An object of class 'ALKmodel'}
}
\value{
  a p-value.
}
\description{
  Compute approximate likelihood ratio test for the
  equality of two age-length keys.
}

