\name{extract-methods}
\docType{methods}
\alias{extract-methods}
\alias{extract,clogit-method}
\alias{extract,ergm-method}
\alias{extract,glm-method}
\alias{extract,gls-method}
\alias{extract,lm-method}
\alias{extract,lme-method}
\alias{extract,lnam-method}
\alias{extract,lrm-method}
\alias{extract,mer-method}
\alias{extract,plm-method}
\alias{extract,pmg-method}
\alias{extract,rq-method}
\alias{extract,systemfit-method}
\title{Methods for Function \code{extract} in Package \pkg{texreg}}
\description{
Methods for function \code{extract} in package \pkg{texreg}.
}
\section{Methods}{
\describe{

\item{\code{extract.clogit}}{
An extract function for clogit objects from the \pkg{survival} package.
}

\item{\code{extract.ergm}}{
An extract function for ergm objects from the \pkg{ergm} package.
}

\item{\code{extract.glm}}{
An extract function for glm objects from the \pkg{stats} package.
}

\item{\code{extract.gls}}{
An extract function for gls objects from the \pkg{nlme} package.
}

\item{\code{extract.lm}}{
An extract function for lm objects from the \pkg{stats} package.
}

\item{\code{extract.lme}}{
An extract function for lme objects from the \pkg{nlme} package.
}

\item{\code{extract.lnam}}{
An extract function for lnam objects from the \pkg{sna} package.
}

\item{\code{extract.lrm}}{
An extract function for lrm objects from the \pkg{Design} or \pkg{rms} package.
}

\item{\code{extract.mer}}{
An extract function for mer objects from the \pkg{lme4} package.
}

\item{\code{extract.plm}}{
An extract function for plm objects from the \pkg{plm} package.
}

\item{\code{extract.pmg}}{
An extract function for pmg objects from the \pkg{plm} package.
}

\item{\code{extract.rq}}{
An extract function for rq objects from the \pkg{quantreg} package.
}

\item{\code{extract.systemfit}}{
An extract function for systemfit objects from the \pkg{systemfit} package.
}
}}
\seealso{
\code{\link{texreg-package} \link{texreg} \link{extract}}
}
\keyword{methods}
