% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackview_functions.R
\name{get_coverage_in_range}
\alias{get_coverage_in_range}
\title{Get Coverage in Range}
\usage{
get_coverage_in_range(bwList, target.range, names, scaling.factor = NULL)
}
\arguments{
\item{bwList}{vector: list of files to use to get coverage. Currently only accepts BigWig files}

\item{target.range}{GRange: specifies the range in which to get coverage}

\item{names}{vector: names to give each GRange coverage created}

\item{scaling.factor}{vector: values by which to scale each coverage value. See get_scaling_factor for more information.}
}
\value{
cvg.L GRangesList: coverage for each file supplied
}
\description{
Get Coverage in Range
}

