% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{get_tx_annotation}
\alias{get_tx_annotation}
\title{get_tx_annotation}
\usage{
get_tx_annotation(range, tx_data, no_introns = FALSE)
}
\arguments{
\item{range}{}

\item{tx_data}{}

\item{no_introns}{}

\item{db_object}{}
}
\description{
get_tx_annotation
}
