% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_range.R
\name{get_view_range}
\alias{get_view_range}
\title{Title}
\usage{
get_view_range(db_object, symbol = NULL, keytype = c("SYMBOL", "TXNAME"),
  chr = NULL, start = NULL, end = NULL, strand = c("*", "+", "-"))
}
\arguments{
\item{db_object}{}

\item{symbol}{}

\item{chr}{}

\item{start}{}

\item{end}{}

\item{strand}{}

\item{tx_data}{}
}
\description{
Title
}
