\name{helpBrowse}
\alias{helpBrowse}
\docType{package}
\title{metrumrg: Extended Examples}
\description{Browse scripts showing use of package tools.}
\details{
\link[=metrumrg]{metrumrg}

I want to ....
\itemize{
\item \href{http://metrumrg.googlecode.com/svn/trunk/inst/example/project/script/assemble.pdf}{assemble a data set}
\item \href{http://metrumrg.googlecode.com/svn/trunk/inst/example/project/script/model.pdf}{run models, make run logs, run predictive checks, launch bootstraps}
\item \href{http://metrumrg.googlecode.com/svn/trunk/inst/example/project/script/covplot.pdf}{plot covariate effects}
\item \href{http://metrumrg.googlecode.com/svn/trunk/inst/example/project/script/parameters.pdf}{generate parameter tables}
\item \href{http://metrumrg.googlecode.com/svn/trunk/inst/example/project/script/simpar.pdf}{simulate with parameter uncertainty}
\item \href{http://metrumrg.googlecode.com/svn/trunk/inst/example/wikimath/wikimath.pdf}{use wikimath markup}
\item \href{http://metrumrg.googlecode.com/svn/trunk/inst/example/tabular/tabular.pdf}{make simple latex tables}
}
}
