\name{helpClasses}
\alias{helpClasses}
\docType{package}
\title{metrumrg: Data Classes}
\description{Classes for building feature-rich data.}
\details{
\link[=metrumrg]{metrumrg} > \link[=helpPrepare]{Data}

I want to ....
\itemize{
\item \link[=as.comment]{create and manipulate comment objects}
\item \link[=as.flag]{create and manipulate flags}
\item \link[=as.keyed]{create and manipulate keyed data.frames}
\item \link[=as.vector.keyed]{column comparisons for keyed data.frames}
\item \link[=as.second]{interconvert time units}
\item \link[=Ops.keyed]{join or subtract keyed data.frames}
\item \link[=as.nm]{create and manipulate nm objects}
\item \link[=as.moot]{modify the assembly of nm objects}
}
}
