\name{helpReport}
\alias{helpReport}
\docType{package}
\title{\link[=MIfuns]{MIfuns}: Reporting}
\description{Tools for summarizing and reporting pharmacometric analyses.}
\details{
I want to ....
\itemize{
\item \link[=lookup]{Lookup} some other form of parameter names
\item \link[=rlog]{Generate} a combined run log from multiple NONMEM runs
}
}
