\name{helpReport}
\alias{helpReport}
\docType{package}
\title{MIfuns: Reporting}
\description{Tools for summarizing and reporting pharmacometric analyses.}
\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item \link[=params]{Identify} documented parameter names
\item \link[=lookup]{Lookup} some other form of parameter names
\item \link[=rlog]{Generate} a combined run log from multiple NONMEM runs
\item \link[=partab]{Create} a model parameter table
}
}
