% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gviz_wrappers.R
\name{make_data_tracks}
\alias{make_data_tracks}
\title{Make Data Tracks}
\usage{
make_data_tracks(cvg_list, gr, genome, chr, bg_title = "black",
  colors = NULL, type = NULL, binwidth = 1000, val = "score",
  hm_scaling = log, kwargs = list())
}
\arguments{
\item{cvg_list}{}

\item{gr}{GRange to plot}

\item{genome}{str}

\item{chr}{str}

\item{bg_title}{str: color for track panel backgrounds}

\item{colors}{list-optional: colors for coverage tracks}

\item{type}{str: 'hist' for multiple coverage tracks, 'heatmap' for condensed view}

\item{binwidth}{scalar}

\item{val}{str: value in cvg.L elements to use for binning coverage. Default is "score"}

\item{kwargs}{}

\item{scaling}{}
}
\value{
dtrack list: DataTrack objects
}
\description{
Builds Gviz DataTrack objects for plotting based on coverage
}
\author{
Justin Finkle
}

