% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browserly_utils.R
\name{make_subplots}
\alias{make_subplots}
\title{make_subplots
Make subplots of coverage tracks}
\usage{
make_subplots(plot_data, type, legend = c("first", "all", "none"), ...)
}
\arguments{
\item{plot_data}{list: a named list. Each list item will be plotted on a 
separate axis, and is expected to be a GRanges object, whose metadata columns
 will be plotted.}

\item{type}{str: type of plot to use}

\item{legend}{str: which subplots to include with legends. 'first' (default), 
only shows the legend for the first subplot.}

\item{...}{additional parameters to pass to browserly_cvg_track}
}
\value{

}
\description{
make_subplots
Make subplots of coverage tracks
}
\author{
Justin Finkle
}
