% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackview_functions.R
\name{make_track_function}
\alias{make_track_function}
\title{Title}
\usage{
make_track_function(cvg_files, genome, db_object, tx_data, cvg_scaling = NULL,
  bg_title = "black", colors = NULL, type = NULL, hm_thresh = 4,
  hm_binsize = 1000, hm_scaling = NULL, snp_gr = NULL, tss_gr = NULL,
  ucsc = FALSE, existing_tracks = NULL, gatrack_kwargs = list(),
  grtrack_kwargs = list(), tsstrack_kwargs = list(),
  snptrack_kwargs = list(), dtrack_kwargs = list(), plot_kwargs = list())
}
\arguments{
\item{cvg_files}{}

\item{genome}{}

\item{db_object}{}

\item{tx_data}{}

\item{cvg_scaling}{}

\item{bg_title}{}

\item{colors}{}

\item{type}{}

\item{hm_thresh}{}

\item{hm_binsize}{}

\item{snp_gr}{}

\item{tss_gr}{}

\item{ucsc}{}

\item{gatrack_kwargs}{}

\item{grtrack_kwargs}{}

\item{tsstrack_kwargs}{}

\item{snptrack_kwargs}{}

\item{dtrack_kwargs}{}

\item{plot_kwargs}{}

\item{target_range}{}
}
\description{
Title
}

