\name{micEcon-internal}
\alias{aidsCoef}
\alias{aidsRestr}
\alias{aidsSystem}
\alias{print.aidsEst}
\alias{snqProfitCoef}
\alias{snqProfitCoefNames}
\alias{snqProfitSystem}
\alias{snqProfitRestrict}


\title{ Internal micEcon Functions }

\description{
   Internal micEcon Functions
}

\details{
   These are not to be called by the user (or in some cases are just
   waiting for proper documentation to be written :).
}

\keyword{ internal }
