% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_heatmaps.R
\name{multi_coverage_heatmap}
\alias{multi_coverage_heatmap}
\title{multi_coverage_heatmap}
\usage{
multi_coverage_heatmap(mats, x = iHeatmapR:::default_x(mats[[1]]),
  y = iHeatmapR:::default_y(mats[[1]]), include = 1000,
  include_method = c("signal", "first", "random"), row_order = c("none",
  "hclust", "kmeans", "groups", "signal"), k = NULL, groups = NULL,
  clust_dist = stats::dist, cluster_by = c("first", "all"),
  signal = lapply(mats, rowMeans, na.rm = TRUE), plot_signal = TRUE,
  name = "Coverage", signal_name = "Avg. Coverage", summary = TRUE,
  source = "HM", scale_method = c("localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar", "none"), pct = 0.95,
  scale_factor = rep(1, length(mats)), share_z = TRUE, ticktext = TRUE,
  start = x[1], end = default_end(x), xlab = "Position",
  font = list(size = 8), ...)
}
\arguments{
\item{x}{x axis labels}

\item{y}{y axis labels}

\item{row_order}{row order method}

\item{k}{k to use for kmeans clustering or cutting heirarchical clustering}

\item{groups}{pre-determined groups for rows}

\item{signal}{signal along row}

\item{name}{name of colorbar}

\item{summary}{make summary plot, boolean}

\item{source}{source name in plotly}

\item{scale_method}{method to use for scaling}

\item{share_z}{share colorbar between heatmaps}

\item{mat}{coverage matrix}

\item{scale}{scale rows? or none}
}
\value{
iHeatmap object
}
\description{
multi_coverage_heatmap
}
\author{
Alicia Schep
}

