% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_single_coverage_heatmap.R
\name{new_single_coverage_heatmap}
\alias{new_single_coverage_heatmap}
\title{new_single_coverage_heatmap}
\usage{
new_single_coverage_heatmap(mat, x = iHeatmapR:::default_x(mat),
  y = iHeatmapR:::default_y(mat), include = 1000,
  include_method = c("signal", "first", "random"), row_order = c("none",
  "hclust", "kmeans", "groups", "signal"), k = NULL, groups = NULL,
  clust_dist = stats::dist, signal = rowSums(mat, na.rm = TRUE),
  plot_signal = TRUE, name = "Coverage", signal_name = "Aggregate",
  summary = TRUE, source = "HM", scale_method = c("localRms", "localMean",
  "localNonZeroMean", "PercentileMax", "scalar", "none"), pct = 0.95,
  scale_factor = 1, ticktext = TRUE, start = x[1], end = default_end(x),
  xlab = "Position", font = list(size = 8), ...)
}
\arguments{
\item{mat}{coverage matrix}

\item{x}{x axis labels}

\item{y}{y axis labels}

\item{row_order}{row order method}

\item{k}{k to use for kmeans clustering or cutting heirarchical clustering}

\item{groups}{pre-determined groups for rows}

\item{signal}{signal along row}

\item{name}{name of colorbar}

\item{summary}{make summary plot, boolean}

\item{source}{source name in plotly}
}
\value{
iHeatmap object
}
\description{
new_single_coverage_heatmap
}
\author{
Alicia Schep
}

