    \name{peakbound}
          \Rdversion{1.0}
          \alias{peakbound}
          \title{Return Exact Peak Bounds}
          \description{
            Imports basecounts from significant windows using generated coordinate file from getsigwindows(), each string of basecounts much be length 1500 (for now)
          }
          \usage{peakbound(pathtocoordfile}
          \arguments{
            \item{pathtocoordfile}{path to coordinate file from getsigwindows()}
          }
          \seealso{
            \code{\link{save}}.
          }
          \examples{

   
          }
          \keyword{file}

