% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_ly.R
\name{plot_browserly_tracks}
\alias{plot_browserly_tracks}
\title{Title}
\usage{
plot_browserly_tracks(txdb, range, tx_data, cvg_list, hm_thresh = 4,
  type = NULL, binsize = 1000)
}
\arguments{
\item{txdb}{}

\item{range}{}

\item{tx_data}{}

\item{cvg_list}{}

\item{hm_thresh}{}

\item{type}{}

\item{binsize}{}
}
\description{
Title
}

