% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser_ly.R
\name{plot_single_locus}
\alias{plot_single_locus}
\title{Title}
\usage{
plot_single_locus(target_range, tx_data, cvg, hm_thresh = 4,
  stacking = "dense", sync_y = TRUE, type = NULL, ...)
}
\arguments{
\item{tx_data}{}

\item{hm_thresh}{}

\item{type}{}

\item{txdb}{}

\item{range}{}
}
\description{
Title
}

