% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackview_functions.R
\name{plot_track_view}
\alias{plot_track_view}
\title{plot genomic coverage along the gene
Adapted from gChipseq function of the same name. Credit to Jinfeng Liu
plot genomic coverage along the gene}
\usage{
plot_track_view(cvg_list, target_range, exon_data, genome, bg_title = "black",
  colors = NULL, type = NULL, hm_thresh = 4, hm_binsize = 1000,
  hm_scaling = NULL, snp_gr = NULL, tss_gr = NULL, ucsc = FALSE,
  gatrack_kwargs = list(), grtrack_kwargs = list(),
  tsstrack_kwargs = list(), snptrack_kwargs = list(),
  dtrack_kwargs = list(), plot_kwargs = list())
}
\arguments{
\item{cvg_list}{GRangesList: typically the results of importing wig/bw/... files for each sample}

\item{target_range}{GRange: range to plot}

\item{exon_data}{data.frame: exon information}

\item{genome}{str: genome build, e.g. hg19, GRCh38, etc.}

\item{bg_title}{}

\item{colors}{str-optional: colors for coverage data tracks}

\item{type}{str-optional: type of datatrack to plot. Currently only 'hist' and 'heatmap' are supported}

\item{hm_thresh}{integer-optional: threshold above which a heatmap is plotted. Can change threshold or override with type}

\item{hm_binsize}{}

\item{snp_gr}{}

\item{tss_gr}{}

\item{ucsc}{}

\item{gatrack_kwargs}{}

\item{grtrack_kwargs}{list-optional: additional arguments for GeneRegionTrack}

\item{tsstrack_kwargs}{}

\item{snptrack_kwargs}{list-optional: additional arguments for the SNP DataTrack}

\item{dtrack_kwargs}{list-optional: additional arguments for DataTrack. These should only be used for global arguments that will apply to all datatracks}

\item{plot_kwargs}{}

\item{scaling}{}
}
\value{
ptracks list: track objects plotted by Gviz
}
\description{
plot genomic coverage along the gene
Adapted from gChipseq function of the same name. Credit to Jinfeng Liu
plot genomic coverage along the gene
}
\author{
Justin Finkle
}

