\name{probTTree}
\alias{probTTree}
\title{Calculates the log-probability of a transmission tree
(assumes w is Gamma(2,1)) #TODO relax this}
\usage{
probTTree(ttree, R, pi)
}
\arguments{
  \item{ttree}{Transmission tree}

  \item{R}{Basic reproduction number}

  \item{pi}{probability of sampling an infected individual}
}
\value{
Probability of the transmission tree
}
\description{
Calculates the log-probability of a transmission tree
(assumes w is Gamma(2,1)) #TODO relax this
}

